% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr_smpdb.R
\name{bitr_smpdb}
\alias{bitr_smpdb}
\title{convert metabolite IDs}
\usage{
bitr_smpdb(mbid, from_Type, to_Type)
}
\arguments{
\item{mbid}{query vector of metabolite ids}

\item{from_Type}{input id type, character}

\item{to_Type}{output id type, character}
}
\value{
a dataframe
}
\description{
You can choose one id type as input and convert to other id type
}
\examples{

bitr_smpdb(c("HMDB0000538","HMDB0000161","HMDB0000045"),
from_Type = "HMDB.ID",to_Type = "ChEBI.ID")

}
