% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichBugSigDB.R
\name{enrichBugSigDB}
\alias{enrichBugSigDB}
\title{BugSigDB enrichment analysis for microbiome data}
\usage{
enrichBugSigDB(
  microbe_list,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  refresh = FALSE
)
}
\arguments{
\item{microbe_list}{a vector of microbe NCBI taxonomy ids.}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none".}

\item{universe}{universe background microbes. If missing, use the BugSigDB
dataset background.}

\item{minGSSize}{minimal size of microbes annotated by a signature term for
testing.}

\item{maxGSSize}{maximal size of microbes annotated by a signature term for
testing.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report.}

\item{refresh}{whether to force a fresh download of the remote artifact.}
}
\value{
An \code{enrichResult} instance.
}
\description{
BugSigDB enrichment analysis for microbiome data
}
\examples{
\dontrun{
taxa <- c("1224", "1236", "91347")
enrichBugSigDB(taxa)
}
}
