% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichHMDB.R
\name{enrichHMDB}
\alias{enrichHMDB}
\title{Metabolism enrichment analysis for microbiome data}
\usage{
enrichHMDB(
  metabo_list,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{metabo_list}{a vector of metabolites in HMDB.ID}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none".}

\item{universe}{universe background genes. If missing, use HMDB db.}

\item{minGSSize}{minimal size of genes annotated by KEGG term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report.}
}
\value{
A \code{enrichResult} instance.
}
\description{
Metabolism enrichment analysis for microbiome data
}
\examples{

x1 <- c("HMDB0000001","HMDB0000005","HMDB0000008")
x2 <- enrichHMDB(x1)

}
