% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMBKEGG.R
\name{enrichMBKEGG}
\alias{enrichMBKEGG}
\title{Metabolism enrichment analysis for microbiome data}
\usage{
enrichMBKEGG(
  metabo_list,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{metabo_list}{a vector of metabolites in KEGG.ID}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none".}

\item{universe}{universe background genes. If missing, use KEGG as default.}

\item{minGSSize}{minimal size of genes annotated by KEGG term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report.}
}
\value{
A \code{enrichResult} instance.
}
\description{
Metabolism enrichment analysis for microbiome data
}
\examples{

mblist3 <- c("C00019","C00020","C00022")
mb3 <- enrichMBKEGG(mblist3)
head(mb3)

}
