% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson_KO.R
\name{gson_cpd}
\alias{gson_cpd}
\title{gson_cpd}
\usage{
gson_cpd()
}
\value{
a 'GSON' object
}
\description{
download compound annotation of the latest version of KEGG pathway and stored in a 'GSON' object
}
\examples{

cpd_gson <- gson_cpd()

}
