% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson_KO.R
\name{gson_enzyme}
\alias{gson_enzyme}
\title{gson_enzyme}
\usage{
gson_enzyme()
}
\value{
a 'GSON' object
}
\description{
download compound annotation of the latest version of KEGG pathway to enzyme and stored in a 'GSON' object
}
