% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson_KO.R
\name{gson_KO}
\alias{gson_KO}
\title{gson_KO}
\usage{
gson_KO()
}
\value{
a 'GSON' object
}
\description{
download KO annotation of the latest version of KEGG pathway and stored in a 'GSON' object
}
\examples{
\dontrun{
ko_gson <- gson_KO()
}

}
