% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson_KO.R
\name{gson_module}
\alias{gson_module}
\title{gson_module}
\usage{
gson_module(db = "ko")
}
\arguments{
\item{db}{ko or enzyme}
}
\value{
a 'GSON' object
}
\description{
download compound annotation of the latest version of KEGG Module and stored in a 'GSON' object
}
