% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gut2014}
\alias{human_gut2014}
\title{A list with 6980 KOs of human gut microbiome}
\format{
A list with KOs of human gut microbiome
}
\usage{
human_gut2014
}
\description{
Li, J. et al. An integrated catalog of reference genes in the human gut microbiome. Nat. Biotechnol. (2014).
}
\details{
data(human_gut2014)
}
\keyword{datasets}
