% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gut2016}
\alias{human_gut2016}
\title{A list with 7189 KOs of human gut microbiome}
\format{
A list with KOs of human gut microbiome
}
\usage{
human_gut2016
}
\description{
Xie, H. et al. Shotgun Metagenomics of 250 Adult Twins Reveals Genetic and Environmental
Impacts on the Gut Microbiome. Cell Syst. 0, 32–46 (2016).
}
\details{
data(human_gut2016)
}
\keyword{datasets}
