% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_skin}
\alias{human_skin}
\title{A list with 10961 KOs of human skin microbiome}
\format{
A list with KOs of human skin microbiome
}
\usage{
human_skin
}
\description{
Li Z , Xia J , Jiang L , et al. Characterization of the human skin resistome and
identification of two microbiota cutotypes[J]. Microbiome, 2021, 9(1).
}
\details{
data(human_skin)
}
\keyword{datasets}
