% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_vagina}
\alias{human_vagina}
\title{A list with 10961 KOs of human vagina microbiome}
\format{
A list with KOs of human vagina microbiome
}
\usage{
human_vagina
}
\description{
Ma B, France MT, Crabtree J, Holm JB, Humphrys MS, Brotman RM, Ravel J. A comprehensive
non-redundant gene catalog reveals extensive within-community intraspecies diversity in
the human vagina. Nat Commun. 2020 Feb 26;11(1):940.
}
\details{
data(human_vagina)
}
\keyword{datasets}
