% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPF_universe}
\alias{IPF_universe}
\title{A list with 2706 universe KOs in IPF patients v.s. healthy individuals.}
\format{
A list with 2706 universe KOs in IPF patients v.s. healthy individuals.
}
\usage{
IPF_universe
}
\description{
Tong, Xunliang et al. “Alterations to the Lung Microbiome in Idiopathic Pulmonary Fibrosis Patients.”
Frontiers in cellular and infection microbiology vol. 9 149. 21 May. 2019, doi:10.3389/fcimb.2019.00149*
}
\details{
data(IPF_universe)
}
\keyword{datasets}
