% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb_examplelist}
\alias{mb_examplelist}
\title{A list of metabolites}
\format{
A list of metabolites
}
\usage{
mb_examplelist
}
\description{
A list of metabolites
}
\details{
data(mb_examplelist)
}
\keyword{datasets}
