% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbiota_data}
\alias{microbiota_data}
\title{54 bacterial genera tested for significantly between T2D metformin samples}
\description{
Reported on Forslund K,et al.2016 (doi: 10.1038/nature15766)
}
\examples{
data(microbiota_data)
}
\keyword{data}
