% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbiota_taxlist}
\alias{microbiota_taxlist}
\alias{microbiota_data}
\title{Example data: a vector of 54 bacterial genera tested for significantly
between T2D metformin samples}
\format{
a vector with 54 genera tax ids
}
\description{
This example data was reported
on Forslund K,et al.2016 (doi: 10.1038/nature15766) and used for
Microbe-Disease Association analysis.
}
\examples{
data(microbiota_taxlist)
}
\references{
\url{https://www.nature.com/articles/nature15766}
}
\keyword{data}
