% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Psoriasis_data}
\alias{Psoriasis_data}
\title{Example data: 134 significantly different functional
COGs between Psoriasis patients and controls}
\value{
\url{https://www.frontiersin.org
/articles/10.3389/fcimb.2021.605825/full}
}
\description{
This example data was
reported on Xiao S,et al.2021 (doi:10.3389/fcimb.2021.605825) and used for
COG enrichment analysis.
}
\examples{
data(Psoriasis_data)
}
\keyword{data}
