% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rat_data}
\alias{Rat_data}
\title{91 KEGG Orthologies (KOs) showing significant associations with weaning weight}
\description{
Reported on Fang S,et al.2019 (doi: 10.1111/1751-7915.13485)
}
\examples{
data(Rat_data)
}
\keyword{data}
