% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr_smpdb.R, R/usingclusterProfiler.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{enricher}
\alias{dotplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{clusterProfiler}{\code{\link[clusterProfiler]{enricher}}}

  \item{enrichplot}{\code{\link[enrichplot]{dotplot}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}
}}

