% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtr.R
\name{taxtr}
\alias{taxtr}
\title{convertion of microne ncbi id and ScientificName}
\usage{
taxtr(Input, Type, Level)
}
\arguments{
\item{Input}{a query vector of microbe ncbi ids or ScientificName}

\item{Type}{The Type of Input, should be "TaxId" or "ScientificName"}

\item{Level}{The taxon level of Input, only "species" and "genus" are accepted.}
}
\value{
a dataframe
}
\description{
convertion of microne ncbi id and ScientificName
}
\examples{
\dontrun{
taxtr(Input = c("2840314","2839514","2839126","2794228"), Type = "TaxId", Level = "genus")
}
}
