% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichEggNOG.R
\name{enrichEggNOG}
\alias{enrichEggNOG}
\title{eggNOG-based KEGG pathway enrichment for microbiome data}
\usage{
enrichEggNOG(
  gene,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  refresh = FALSE
)
}
\arguments{
\item{gene}{a vector of eggNOG orthologous group identifiers.}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none".}

\item{universe}{background eggNOG orthologous groups. If missing, use the
eggNOG dataset background.}

\item{minGSSize}{minimal size of pathway gene sets for testing.}

\item{maxGSSize}{maximal size of pathway gene sets for testing.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report.}

\item{refresh}{whether to force a fresh download of the remote eggNOG
artifact.}
}
\value{
An \code{enrichResult} instance.
}
\description{
eggNOG-based KEGG pathway enrichment for microbiome data
}
\examples{
\dontrun{
eggnog_obj <- MicrobiomeProfiler:::mp_eggnog_gson()
gsid2gene <- methods::slot(eggnog_obj, "gsid2gene")
pathway_sizes <- sort(table(gsid2gene$gsid))
pathway_id <- names(pathway_sizes[pathway_sizes >= 3])[1]
og <- head(unique(gsid2gene[gsid2gene$gsid == pathway_id, "gene"]), 3)
enrichEggNOG(og, minGSSize = 1)
}
}
