% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichEggNOG.R
\name{gseEggNOG}
\alias{gseEggNOG}
\title{GSEA with eggNOG-based KEGG pathway annotations}
\usage{
gseEggNOG(
  geneList,
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  refresh = FALSE
)
}
\arguments{
\item{geneList}{order ranked geneList.}

\item{exponent}{weight of each step.}

\item{minGSSize}{minimal size of each geneSet for analyzing.}

\item{maxGSSize}{maximal size of genes annotated for testing.}

\item{pvalueCutoff}{pvalue Cutoff.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none".}

\item{refresh}{whether to force a fresh download of the remote eggNOG
artifact.}
}
\value{
A \code{gseaResult} instance.
}
\description{
GSEA with eggNOG-based KEGG pathway annotations
}
\examples{
\dontrun{
geneList <- c(2.5, 1.2, -1.8)
names(geneList) <- c("OG0001", "OG0002", "OG0003")
gseEggNOG(geneList, minGSSize = 1)
}
}
