% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeProfiler-package.R
\docType{package}
\name{MicrobiomeProfiler-package}
\alias{MicrobiomeProfiler-package}
\alias{MicrobiomeProfiler}
\title{Functional enrichment analysis for microbiome data}
\description{
The package implements an shiny application for functional enrichment
analysis and visualiztion of microbiome studies.
}
