% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Psoriasis_data}
\alias{Psoriasis_data}
\title{134 significantly different functional COGs between Psoriasis patients and controls}
\description{
Reported on Xiao S,et al.2021 (doi:10.3389/fcimb.2021.605825)
}
\examples{
data(Psoriasis_data)
}
\keyword{data}
