% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rat_data}
\alias{Rat_data}
\title{Example data: a vector of 91 KEGG Orthologies (KOs) showing
significant associations with weaning weight}
\format{
a vector of 91 KEGG Orthologies (KOs)
}
\description{
This example data was reported on
Fang S,et al.2019 (doi: 10.1111/1751-7915.13485) and used for KEGG
enrichment analysis.
}
\examples{
data(Rat_data)
}
\references{
\url{https://sfamjournals.onlinelibrary.wiley.com
/doi/10.1111/1751-7915.13485}
}
\keyword{data}
